Make tests for app and plugin generators more DRY
