+++ railties/test/fixtures/lib/app_builders/empty_builder.rb
