+++ railties/test/fixtures/lib/app_builders/simple_builder.rb
@@ -1,0 +1,7 @@
+class AppBuilder
+  def gitignore
+    create_file ".gitignore", <<-R.strip
+foobar
+    R
+  end
+end
