+++ railties/test/fixtures/lib/app_builders/tweak_builder.rb
@@ -1,0 +1,7 @@
+class AppBuilder < Rails::AppBuilder
+  def gitignore
+    create_file ".gitignore", <<-R.strip
+foobar
+    R
+  end
+end
