--- railties/test/fixtures/lib/tweak_builder.rb
@@ -1,7 +1,0 @@
-class AppBuilder < Rails::AppBuilder
-  def configru
-    create_file "config.ru", <<-R.strip
-run proc { |env| [200, { "Content-Type" => "text/html" }, ["Hello World"]] }
-    R
-  end
-end
