+++ railties/test/generators/app_generator_test.rb
@@ -1,6 +1,7 @@
 require 'abstract_unit'
 require 'generators/generators_test_helper'
 require 'rails/generators/rails/app/app_generator'
+require 'generators/shared_generator_tests.rb'
 
 DEFAULT_APP_FILES = %w(
   .gitignore
@@ -40,55 +41,16 @@ DEFAULT_APP_FILES = %w(
 class AppGeneratorTest < Rails::Generators::TestCase
   include GeneratorsTestHelper
   arguments [destination_root]
+  include SharedGeneratorTests
 
-  def setup
-    Rails.application = TestApp::Application
-    super
-    Rails::Generators::AppGenerator.instance_variable_set('@desc', nil)
-    @bundle_command = File.basename(Thor::Util.ruby_command).sub(/ruby/, 'bundle')
-    
-    Kernel::silence_warnings do
-      Thor::Base.shell.send(:attr_accessor, :always_force)
-      @shell = Thor::Base.shell.new
-      @shell.send(:always_force=, true)
-    end
-  end
-
-  def teardown
-    super
-    Rails::Generators::AppGenerator.instance_variable_set('@desc', nil)
-    Rails.application = TestApp::Application.instance
-  end
-
-  def test_application_skeleton_is_created
-    run_generator
-
-    DEFAULT_APP_FILES.each{ |path| assert_file path }
-  end
-
-  def test_application_generate_pretend
-    run_generator ["testapp", "--pretend"]
-
-    DEFAULT_APP_FILES.each{ |path| assert_no_file path }
+  def default_files
+    ::DEFAULT_APP_FILES
   end
 
   def test_application_controller_and_layout_files
     run_generator
     assert_file "app/views/layouts/application.html.erb", /stylesheet_link_tag :all/
     assert_no_file "public/stylesheets/application.css"
-  end
-
-  def test_options_before_application_name_raises_an_error
-    content = capture(:stderr){ run_generator(["--skip-active-record", destination_root]) }
-    assert_equal "Options should be given after the application name. For details run: rails --help\n", content
-  end
-
-  def test_name_collision_raises_an_error
-    reserved_words = %w[application destroy plugin runner test]
-    reserved_words.each do |reserved|
-      content = capture(:stderr){ run_generator [File.join(destination_root, reserved)] }
-      assert_equal "Invalid application name #{reserved}. Please give a name which does not match one of the reserved rails words.\n", content
-    end
   end
 
   def test_invalid_database_option_raises_an_error
@@ -99,13 +61,6 @@ class AppGeneratorTest < Rails::Generators::TestCase
   def test_invalid_application_name_raises_an_error
     content = capture(:stderr){ run_generator [File.join(destination_root, "43-things")] }
     assert_equal "Invalid application name 43-things. Please give a name which does not start with numbers.\n", content
-  end
-
-  def test_application_name_raises_an_error_if_name_already_used_constant
-    %w{ String Hash Class Module Set Symbol }.each do |ruby_class|
-      content = capture(:stderr){ run_generator [File.join(destination_root, ruby_class)] }
-      assert_equal "Invalid application name #{ruby_class}, constant #{ruby_class} is already in use. Please choose another application name.\n", content
-    end
   end
 
   def test_invalid_application_name_is_fixed
@@ -217,34 +172,9 @@ class AppGeneratorTest < Rails::Generators::TestCase
     assert_file "public/javascripts/rails.js", /jQuery/
   end
 
-  def test_shebang_is_added_to_rails_file
-    run_generator [destination_root, "--ruby", "foo/bar/baz"]
-    assert_file "script/rails", /#!foo\/bar\/baz/
-  end
-
-  def test_shebang_when_is_the_same_as_default_use_env
-    run_generator [destination_root, "--ruby", Thor::Util.ruby_command]
-    assert_file "script/rails", /#!\/usr\/bin\/env/
-  end
-
   def test_template_from_dir_pwd
     FileUtils.cd(Rails.root)
     assert_match /It works from file!/, run_generator([destination_root, "-m", "lib/template.rb"])
-  end
-
-  def test_template_raises_an_error_with_invalid_path
-    content = capture(:stderr){ run_generator([destination_root, "-m", "non/existant/path"]) }
-    assert_match /The template \[.*\] could not be loaded/, content
-    assert_match /non\/existant\/path/, content
-  end
-
-  def test_template_is_executed_when_supplied
-    path = "http://gist.github.com/103208.txt"
-    template = %{ say "It works!" }
-    template.instance_eval "def read; self; end" # Make the string respond to read
-
-    generator([destination_root], :template => path).expects(:open).with(path, 'Accept' => 'application/x-thor-template').returns(template)
-    assert_match /It works!/, silence(:stdout){ generator.invoke_all }
   end
 
   def test_usage_read_from_file
@@ -266,19 +196,6 @@ class AppGeneratorTest < Rails::Generators::TestCase
     assert_file 'lib/test_file.rb', 'heres test data'
   end
 
-  def test_dev_option
-    generator([destination_root], :dev => true).expects(:run).with("#{@bundle_command} install")
-    silence(:stdout){ generator.invoke_all }
-    rails_path = File.expand_path('../../..', Rails.root)
-    assert_file 'Gemfile', /^gem\s+["']rails["'],\s+:path\s+=>\s+["']#{Regexp.escape(rails_path)}["']$/
-  end
-
-  def test_edge_option
-    generator([destination_root], :edge => true).expects(:run).with("#{@bundle_command} install")
-    silence(:stdout){ generator.invoke_all }
-    assert_file 'Gemfile', /^gem\s+["']rails["'],\s+:git\s+=>\s+["']#{Regexp.escape("git://github.com/rails/rails.git")}["']$/
-  end
-
 protected
 
   def action(*args, &block)
@@ -292,61 +209,20 @@ class CustomAppGeneratorTest < Rails::Generators::TestCase
   tests Rails::Generators::AppGenerator
 
   arguments [destination_root]
-
-  def setup
-    Rails.application = TestApp::Application
-    super
-    Rails::Generators::AppGenerator.instance_variable_set('@desc', nil)
-    @bundle_command = File.basename(Thor::Util.ruby_command).sub(/ruby/, 'bundle')
-  end
-
-  def teardown
-    super
-    Rails::Generators::AppGenerator.instance_variable_set('@desc', nil)
-    Object.class_eval { remove_const :AppBuilder if const_defined?(:AppBuilder) }
-    Rails.application = TestApp::Application.instance
-  end
-
-  def test_builder_option_with_empty_app_builder
-    FileUtils.cd(Rails.root)
-    run_generator([destination_root, "-b", "#{Rails.root}/lib/empty_builder.rb"])
-    DEFAULT_APP_FILES.each{ |path| assert_no_file path }
-  end
-
-  def test_builder_option_with_simple_app_builder
-    FileUtils.cd(Rails.root)
-    run_generator([destination_root, "-b", "#{Rails.root}/lib/simple_builder.rb"])
-    (DEFAULT_APP_FILES - ['config.ru']).each{ |path| assert_no_file path }
-    assert_file "config.ru", %[run proc { |env| [200, { "Content-Type" => "text/html" }, ["Hello World"]] }]
-  end
-
-  def test_builder_option_with_relative_path
-    here = File.expand_path(File.dirname(__FILE__))
-    FileUtils.cd(here)
-    run_generator([destination_root, "-b", "../fixtures/lib/simple_builder.rb"])
-    (DEFAULT_APP_FILES - ['config.ru']).each{ |path| assert_no_file path }
-    assert_file "config.ru", %[run proc { |env| [200, { "Content-Type" => "text/html" }, ["Hello World"]] }]
-  end
-
-  def test_builder_option_with_tweak_app_builder
-    FileUtils.cd(Rails.root)
-    run_generator([destination_root, "-b", "#{Rails.root}/lib/tweak_builder.rb"])
-    DEFAULT_APP_FILES.each{ |path| assert_file path }
-    assert_file "config.ru", %[run proc { |env| [200, { "Content-Type" => "text/html" }, ["Hello World"]] }]
-  end
-
-  def test_builder_option_with_http
-    path = "http://gist.github.com/103208.txt"
-    template = "class AppBuilder; end"
-    template.instance_eval "def read; self; end" # Make the string respond to read
-
-    generator([destination_root], :builder => path).expects(:open).with(path, 'Accept' => 'application/x-thor-template').returns(template)
-    capture(:stdout) { generator.invoke_all }
-
-    DEFAULT_APP_FILES.each{ |path| assert_no_file path }
-  end
+  include SharedCustomGeneratorTests
 
 protected
+  def default_files
+    ::DEFAULT_APP_FILES
+  end
+
+  def builders_dir
+    "app_builders"
+  end
+
+  def builder_class
+    :AppBuilder
+  end
 
   def action(*args, &block)
     silence(:stdout){ generator.send(*args, &block) }
